# docs/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "DisplayCurrency": {
        "enum": [
          "USD",
          "EUR"
        ],
        "title": "DisplayCurrency",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderDetailRead": {
        "properties": {
          "admin_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Notes"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "client": {
            "$ref": "#/components/schemas/ClientSummary"
          },
          "client_id": {
            "format": "uuid",
            "title": "Client Id",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discount_amount_usd": {
            "title": "Discount Amount Usd",
            "type": "number"
          },
          "discount_percent": {
            "title": "Discount Percent",
            "type": "integer"
          },
          "display_currency": {
            "$ref": "#/components/schemas/DisplayCurrency"
          },
          "final_total_usd": {
            "title": "Final Total Usd",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "items_count": {
            "default": 0,
            "title": "Items Count",
            "type": "integer"
          },
          "order_number": {
            "title": "Order Number",
            "type": "string"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "payment_method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "refunded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refunded At"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "subtotal_usd": {
            "title": "Subtotal Usd",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order_number",
          "client_id",
          "client",
          "status",
          "payment_method",
          "display_currency",
          "subtotal_usd",
          "discount_amount_usd",
          "discount_percent",
          "final_total_usd",
          "comment",
          "admin_notes",
          "paid_at",
          "completed_at",
          "cancelled_at",
          "refunded_at",
          "created_at",
          "updated_at",
          "items"
        ],
        "title": "OrderDetailRead",
        "type": "object"
      },
...
      "OrderStatus": {
        "enum": [
          "pending",
          "paid",
          "in_progress",
          "completed",
          "cancelled",
          "refunded"
        ],
        "title": "OrderStatus",
        "type": "string"
      },
      "OrderStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "OrderStatusUpdate",
        "type": "object"
      },
      "OrderUpdate": {
        "properties": {
          "admin_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Notes"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "title": "OrderUpdate",
        "type": "object"
      },
...
      "Paginated_OrderRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "Paginated[OrderRead]",
        "type": "object"
      },
...
      "PaymentClaimResponse": {
        "description": "Returned by POST /public/orders/{number}/claim-payment.\n\nIdempotent: if the claim was already filed, the original\n`payment_claimed_at` is returned and no new Telegram alert is sent.\nStatus stays `pending` \u2014 only the admin's manual verification\nadvances it to `paid`.",
        "properties": {
          "order_number": {
            "title": "Order Number",
            "type": "string"
          },
          "payment_claimed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Claimed At"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        },
        "required": [
          "order_number",
          "status"
        ],
        "title": "PaymentClaimResponse",
        "type": "object"
      },
      "PaymentMethod": {
        "enum": [
          "paypal",
          "usdt_trc20",
          "card_ecomtrade24"
        ],
        "title": "PaymentMethod",
        "type": "string"
      },
...
      "PublicOrderCreate": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "discord": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discord"
          },
          "display_currency": {
            "$ref": "#/components/schemas/DisplayCurrency",
            "default": "USD"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicOrderItemCreate"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "payment_method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "telegram": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram"
          },
          "whatsapp": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp"
          }
        },
        "required": [
          "email",
          "payment_method",
          "items"
        ],
        "title": "PublicOrderCreate",
        "type": "object"
      },
...
      "PublicOrderResponse": {
        "properties": {
          "checkout_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discount_amount_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Amount Usd"
          },
          "display_currency": {
            "$ref": "#/components/schemas/DisplayCurrency"
          },
          "final_total_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Total Eur"
          },
          "final_total_usd": {
            "title": "Final Total Usd",
            "type": "number"
          },
          "order_number": {
            "title": "Order Number",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        },
        "required": [
          "order_number",
          "status",
          "final_total_usd",
          "display_currency",
          "created_at"
        ],
        "title": "PublicOrderResponse",
        "type": "object"
      },
      "PublicOrderStatusResponse": {
        "description": "Polled by the public payment-success page. Intentionally PII-free \u2014\nanyone with the order_number can read this, same trust level as a\nStripe/PayPal session reference.",
        "properties": {
          "display_currency": {
            "$ref": "#/components/schemas/DisplayCurrency"
          },
          "final_total_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Total Eur"
          },
          "final_total_usd": {
            "title": "Final Total Usd",
            "type": "number"
          },
          "last_updated_at": {
            "format": "date-time",
            "title": "Last Updated At",
            "type": "string"
          },
          "order_number": {
            "title": "Order Number",
            "type": "string"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        },
        "required": [
          "order_number",
          "status",
          "final_total_usd",
          "display_currency",
          "last_updated_at"
        ],
        "title": "PublicOrderStatusResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "REST API for Nanoboost public site and admin panel.",
    "title": "Nanoboost Admin API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/clients/{client_id}/orders": {
      "get": {
        "operationId": "list_client_orders_api_v1_clients__client_id__orders_get",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Client Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_OrderRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Client Orders",
        "tags": [
          "clients"
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "operationId": "list_orders_api_v1_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrderStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "client_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            }
          },
          {
            "in": "query",
            "name": "payment_method",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PaymentMethod"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Payment Method"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_OrderRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Orders",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/orders/{order_id}": {
      "get": {
        "operationId": "get_order_api_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Order",
        "tags": [
          "orders"
        ]
      },
      "patch": {
        "operationId": "update_order_api_v1_orders__order_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Order",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{order_id}/status": {
      "patch": {
        "operationId": "change_order_status_api_v1_orders__order_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Change Order Status",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/public/orders": {
      "post": {
        "operationId": "create_public_order_api_v1_public_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicOrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicOrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Public Order",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/orders/{order_number}/claim-payment": {
      "post": {
        "description": "Customer signals \"I have paid\" for a manual-payment order.\n\nSets `payment_claimed_at` and pings the admin via Telegram so they\ncan verify the wallet/PayPal balance and flip status \u2192 PAID.\n\nIdempotent \u2014 replaying the call returns the original timestamp and\nskips the notification. Only PENDING orders accept new claims;\nterminal-state orders return the current state unchanged.\n\nPublic surface, no auth \u2014 `order_number` is the credential, same\ntrust posture as the polling endpoint.",
        "operationId": "claim_payment_api_v1_public_orders__order_number__claim_payment_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_number",
            "required": true,
            "schema": {
              "title": "Order Number",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Payment",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/public/orders/{order_number}/status": {
      "get": {
        "description": "Polled by the payment-success page after the customer returns from\nthe hosted checkout. Public \u2014 order_number is the only credential, same\ntrust posture as a Stripe/PayPal session reference. PII-free response.",
        "operationId": "get_public_order_status_api_v1_public_orders__order_number__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_number",
            "required": true,
            "schema": {
              "title": "Order Number",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicOrderStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Public Order Status",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}